Move serving plugins and related code to the same directory
