+++ lib/galaxy/visualization/plugins/pluginframework.py
