+++ lib/galaxy/visualization/plugins/registry.py
@@ -10,7 +10,7 @@ import weakref
 from galaxy.web import url_for
 import galaxy.exceptions
 
-from galaxy.web.base import pluginframework
+from galaxy.visualization.plugins import pluginframework
 from galaxy.visualization.plugins import config_parser
 from galaxy.visualization.plugins import plugin as vis_plugins
 from galaxy.visualization.plugins import utils as vis_utils
